0.1.48
none

+++ unauticnainfo/Buttonresset.cs
@@ -1,5 +1,25 @@
 using Godot;
 using System;
+public class GetUniverse
+{
+	static public string Universe()
+	{
+		string universe = "MU Unuaticna";
+		if (System.IO.File.Exists("Unauticna.exe"))
+		{
+			universe = "U Unauticna";
+		}
+		if (System.IO.File.Exists("Deltafate.exe"))
+		{
+			universe = "U Deltafate";
+		}
+		if (System.IO.File.Exists("Enteria.exe"))
+		{
+			universe = "U Enteria";
+		}
+		return universe;
+	}
+}
 
 public class Buttonresset : Button
 {
@@ -15,29 +35,52 @@ public class Buttonresset : Button
 
 	private void _on_Button2_button_down()
 	{
-		if (System.IO.Directory.Exists("unsave"))
+		if (GetUniverse.Universe() == "U Unauticna")
 		{
-			System.IO.Directory.Delete("unsave", true);
+
+			if (System.IO.Directory.Exists("unsave"))
+			{
+				System.IO.Directory.Delete("unsave", true);
+			}
+			if (System.IO.Directory.Exists("munsave"))
+			{
+				System.IO.Directory.Delete("munsave", true);
+			}
+			if (System.IO.Directory.Exists("unsavet"))
+			{
+				System.IO.Directory.Delete("unsavet", true);
+			}
+			if (System.IO.Directory.Exists("debug"))
+			{
+				System.IO.Directory.Delete("debug", true);
+			}
+			if (System.IO.Directory.Exists("world"))
+			{
+				System.IO.Directory.Delete("world", true);
+
+			}
+
 		}
-		if (System.IO.Directory.Exists("munsave"))
+		if (GetUniverse.Universe() == "U Deltafate")
 		{
-			System.IO.Directory.Delete("munsave", true);
-		}
-		if (System.IO.Directory.Exists("unsavet"))
-		{
-			System.IO.Directory.Delete("unsavet", true);
-		}
-		if (System.IO.Directory.Exists("debug"))
-		{
-			System.IO.Directory.Delete("debug", true);
-		}
-		if (System.IO.Directory.Exists("world"))
-		{
-			System.IO.Directory.Delete("world", true);
+
+
+			if (System.IO.Directory.Exists("debug"))
+			{
+				System.IO.Directory.Delete("debug", true);
+			}
+			if (System.IO.Directory.Exists("DELTAFATE"))
+			{
+				System.IO.Directory.Delete("DELTAFATE", true);
+			}
+			if (System.IO.Directory.Exists("sins"))
+			{
+				System.IO.Directory.Delete("sins", true);
+			}
+
 
 		}
 	}
-
-	}
+}
